+++ src/commands/init/index.ts
@@ -1,3 +1,4 @@
+import { confirm } from '@inquirer/prompts'
 import { isSupportedChain, toSupportedChainID } from '@utils/chain'
 import { trim } from '@utils/string'
 
@@ -76,6 +77,18 @@ export const handleInitializeCli = async (cmdInput: InitProjectInput) => {
     }
 
     /**
+     * Analytics
+     */
+    let allowAnalytics = false
+
+    if (!cmdInput.ignoreAnalytics) {
+      allowAnalytics = await confirm({
+        message: `Do you allow INFINIT CLI to send usage data to help improve the tool? (This can be changed later in the config file)`,
+        default: true,
+      })
+    }
+
+    /**
      * Package Manager
      */
 
@@ -85,11 +98,11 @@ export const handleInitializeCli = async (cmdInput: InitProjectInput) => {
      * Initialize CLI Project
      */
 
-    await initializeCliProject(projectDirectory, protocolModule, chainId, packageManager, deployerId)
+    const { generatedScriptFile } = await initializeCliProject(projectDirectory, protocolModule, chainId, packageManager, deployerId, allowAnalytics)
 
     await compileProject(projectDirectory, protocolModule)
 
-    console.log(chalkSuccess(`🔥 Successfully initialized a project, go to ${chalkInfo(`src/scripts/default.script.ts`)} to start building.`))
+    console.log(chalkSuccess(`🔥 Successfully initialized a project, go to ${chalkInfo(`src/scripts/${generatedScriptFile}`)} to start building.`))
   } catch (error) {
     console.error(chalkError(error))
   }